// backend/src/modules/episode/type.ts
export interface CrewMember {
  job: string;
  department: string;
  credit_id: string;
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string;
}

export interface GuestStarDetails {
  character: string;
  credit_id: string;
  order: number;
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string;
}
export interface CastDetails {
  character: string;
  credit_id: string;
  order: number;
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string;
}

export interface EpisodeDetails {
  air_date: string;
  crew: CrewMember[];
  episode_number: number;
  guest_stars: GuestStarDetails[];
  name: string;
  overview: string;
  id: number;
  production_code: string;
  runtime: number;
  season_number: number;
  still_path: string;
  vote_average: number;
  vote_count: number;
}

export interface creditsResponse {
  id: number;
  cast: CastDetails[];
  crew: CrewMember[];
  guest_stars: GuestStarDetails[];
}

export interface ImageDetails {
  aspect_ratio: number;
  height: number;
  iso_639_1?: string;
  file_path: string;
  vote_average: number;
  vote_count: number;
  width: number;
}

export interface GetImagesResponse {
  id: number;
  backdrops: ImageDetails[];
  logos: ImageDetails[];
  posters: ImageDetails[];
}